+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 0.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Migrate away from requireing BASE_URL to be set and just use flask's own url handling [halkeye]
 
 
 0.0.1 (2017-06-21)